+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Additionally output plannotate annotations as a GenBank (`.gbk`) file.
 ### Removed
 - Remove plasmid length column from plannotate feature table.
 ### Changed